app/requirements.txt
streamlit==1.33.0
langchain==0.1.15
duckdb==0.10.1
streamlit_code_editor==0.1.12